Added a TabRender component to reduce redundancy of  of the tabs, fixed the connection to firebase and implemented useEffect hooks to prevent the app from constantly taking data from the databse
